+++ CHANGELOG.md
@@ -10,7 +10,9 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
-- Fix bug in parsing of heading.
+## [0.4.3] - 2021-07-24
+
+- [Fix bug in parsing of heading.](https://github.com/taiki-e/parse-changelog/pull/13)
 
 ## [0.4.2] - 2021-07-22
 
@@ -86,7 +88,8 @@ No public API changes from 0.2.1.
 
 Initial release
 
-[Unreleased]: https://github.com/taiki-e/parse-changelog/compare/v0.4.2...HEAD
+[Unreleased]: https://github.com/taiki-e/parse-changelog/compare/v0.4.3...HEAD
+[0.4.3]: https://github.com/taiki-e/parse-changelog/compare/v0.4.2...v0.4.3
 [0.4.2]: https://github.com/taiki-e/parse-changelog/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/taiki-e/parse-changelog/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/taiki-e/parse-changelog/compare/v0.3.0...v0.4.0